refactor: Move PackageFactory to AmaiSosu.Common namespace
